Hlhome page notifications

+++ www/html/php/disable_chamber.php
@@ -2,11 +2,16 @@
 include 'db_handler.php';
 $db = new DB_Handler();
 
-$chamber = $_POST["chamber"];
+if(isset($_SESSION['user'])){
+    $chamber = $_POST["chamber"];
 
-$results = $db->disableChamber($chamber);
-$results = $db->removeParent($chamber);
+    $results = $db->disableChamber($chamber);
+    $results = $db->removeParent($chamber);
 
-echo json_encode($results);
+    echo json_encode($results);
+}
+else {
+    echo json_encode(false);
+}
 
 ?>
